+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/ManualSessionScenario.m
@@ -6,10 +6,17 @@
 #import "ManualSessionScenario.h"
 #import <Bugsnag/Bugsnag.h>
 
+@interface BugsnagConfiguration ()
+- (void)deletePersistedUserData;
+@end
+
 @implementation ManualSessionScenario
 
 - (void)startBugsnag {
     self.config.autoTrackSessions = NO;
+    self.config.persistUser = NO;
+    [self.config deletePersistedUserData];
+    [self.config setCurrentUser:nil];
     [super startBugsnag];
 }